Crate verify_headers

Utility program to verify properties of headers sequenced by HotShot.

Structs

Options 🔒

Utility program to verify properties of headers sequenced by HotShot.

Functions

get_header 🔒

get_l1_block 🔒

main 🔒

verify_header 🔒

Type Aliases

SequencerClient 🔒